% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{aci_curve}
\alias{aci_curve}
\title{ACI Curve}
\usage{
aci_curve(data, include_TPU_curve = FALSE, C_c_denom_multiplier = 4,
  G_star_denom_multiplier = 8, V_cmax_var = V_cmax, C_c_var = C_c,
  G_star_var = G_star, K_c_var = K_c, O_var = O, K_o_var = K_o,
  R_d_var = R_d, J_var = J, TPU_var = TPU)
}
\arguments{
\item{data}{A dataframe containing all the relevant parameters.}

\item{TPU}{Boolean indicating the inclusion of the TPU limiting curve
(default is FALSE)}
}
\description{
Fits the ACI curve using FvCB model with the option to include
the TPU limiting curve. Currently not working, merely a placeholder.
}
