% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{aci_formula}
\alias{aci_formula}
\title{ACI Formula}
\usage{
aci_formula(V_cmax, C_c, G_star, K_c, O, K_o, R_d, J, TPU_var = NA,
  C_c_denom_multiplier = 4, G_star_denom_multiplier = 8)
}
\arguments{
\item{TPU}{Boolean indicating the inclusion of the TPU limiting curve
(default is FALSE)}
}
\description{
Fits the ACI curve using FvCB model with the option to include
the TPU limiting curve. Currently not working, merely a placeholder.
}
