% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{nlme_aci}
\alias{nlme_aci}
\title{Non-Linear Mixed Effects ACI model}
\usage{
nlme_aci(data, fixed, random, V_cmax, C_c, G_star, K_c, O, K_o, R_d, J,
  TPU_var = NA, C_c_denom_multiplier = 4, G_star_denom_multiplier = 8,
  starting_values = NA, cov_matrix = pdSymm, ...)
}
\arguments{
\item{data}{The dataframe containing photosynthesis data}
}
\value{
An nlme model
}
\description{
Fits a non-linear mixed effects model using \code{nlme} package.
Currently not working, decent placeholder for final version.
}
