% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{non_rectangular_hyperbola}
\alias{non_rectangular_hyperbola}
\title{Non rectangular hyperbola (from Gerrit)}
\usage{
non_rectangular_hyperbola(PAR, J_max, alpha, theta)
}
\arguments{
\item{PAR}{I think this is the assimilation rate?}

\item{alpha}{Just a parameter}

\item{theta}{Ditto}

\item{Jmax}{Something to do with electrons}
}
\description{
Returns the a fit for a non-rectangular hyperbola curve
}
