% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{Rubisco_limited}
\alias{Rubisco_limited}
\title{Rubisco limited curve}
\usage{
Rubisco_limited(V_cmax, C_c, G_star, K_c, O, K_o, R_d)
}
\arguments{
\item{V_cmax}{(\eqn{V_{cmax}}) is the maximum velocity of Rubisco for
carboxylation.}

\item{C_c}{(\eqn{C_C}) is the \eqn{CO_2} partial pressure at Rubisco (within
chloroplast).}

\item{G_star}{(\eqn{\Gamma^*}) is the \eqn{CO_2} at which oxygenation proceeds at
twice the rate of carboxylaiton causing photosynthetic uptake of \eqn{CO_2} to be
exactly compensated by photorespiratory \eqn{CO_2} release.}

\item{K_c}{(\eqn{K_C}) is the Michaelis constant of Rubisco for carbon dioxide.}

\item{O}{is the partial pressure of oxygen at Rubisco.}

\item{K_o}{(\eqn{K_O}) is the inhibition constant of Rubisco for oxygen.}

\item{R_d}{(\eqn{R_d}) is respiratory \eqn{CO_2} release by methods other
than by photorespiration and is presumed to be primarily mitochondiral
respiraion.}
}
\value{
Nothing at all.
}
\description{
Returns the expected CO2 assimilation for the curve restricted by Rubisco
assuming a saturating supply of RuBP. Limitation by Rubisco is associated
with low CO2 rather than V_max. This is part of FvCB model.
}
\examples{
Rubisco_limited(<examples>)
More examples and whatnot
}
