% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{RuBP_limited}
\alias{RuBP_limited}
\title{RuBP limited curve}
\usage{
RuBP_limited(J, C_c, G_star, R_d, C_c_denom_multiplier = 4,
  G_star_denom_multiplier = 8)
}
\arguments{
\item{J}{provided here is the rate of electron transport going to support
NADP+ reduction for RuBP regeneration at the measurement light intensity. J
is sometimes used to estimate a maximum rate that could be obtained at
saturating light, and this is called \eqn{J_{max}}.}

\item{C_c}{(\eqn{C_c}) is the \eqn{CO_2} partial pressure at Rubisco (within
chloroplast).}

\item{G_star}{(\eqn{\Gamma^*}) is the \eqn{CO_2} at which oxygenation proceeds at
twice the rate of carboxylaiton causing photosynthetic uptake of \eqn{CO_2} to be
exactly compensated by photorespiratory \eqn{CO_2} release.}
}
\value{
Nothing at all.
}
\description{
Returns the expected CO2 assimilation for the curve restricted
by Ribulose 1,5-bisphosphate (RuBP) regeneration. This equation assumes
four electrons per carboxylation and oxygenation. Based on the number of
electrons required for NADP+ reduction the conservative values of 4 and 8
are used as defaults for the \eqn{C_C} and \eqn{\Gamma^*} multipliers in the
denominator, but 4.5 and 10.5 have been used.
}
\examples{
RuBP_limited(<examples>)
More examples and whatnot
}
