% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis_curves.R
\name{TPU_limited}
\alias{TPU_limited}
\title{TPU limited curve}
\usage{
TPU_limited(TPU, R_d)
}
\arguments{
\item{TPU}{is the rate of use of triose phosphates but can also be any export
of carbon from the Calvin cycle including direct use of photorespiratory
glycine or serine.}

\item{R_d}{(
%(\eqn{R_d})%
) is respiratory \eqn{CO_2} release by methods other than by
photorespiration and is presumed to be primarily mitochondiral respiraion.}
}
\value{
Nothing at all.
}
\description{
Returns the expected CO2 assimilation when the rate is limited
by Triose Phosphate Use (TPU).
}
\examples{
TPU_limited(<examples>)
More examples and whatnot
}
